requirements.txt
django==3.2.6
gunicorn==20.0.4
psycopg2-binary==2.9.1
gunicorn==20.0.4
elasticsearch==7.15.0
pydantic==1.8.2
python-dateutil==2.8.2
isort==5.10.1